desktop-sandbox/cloudflare/files/agent-smith-py/pyproject.toml
[project]
name = "agent-smith-py"
version = "0.1.0"
description = "Multi-agent workforce using CAMEL-AI"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "camel-ai[all]>=0.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.0.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
browser = [
    "playwright>=1.40.0",
]
full = [
    "playwright>=1.40.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.0",
    "python-docx>=1.0.0",
]

[project.scripts]
agent-smith = "main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src", "src.agents"]
py-modules = ["main"]
